requirements.txt
antares-client==0.3.1
aotools==1.0
appnope==0.1.0
asdf==2.6.0
astroplan==0.6
astropy==4.0
astroquery==0.4
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
bson==0.5.8
certifi==2019.9.11
chardet==3.0.4
Click==7.0

cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.8
django-bootstrap4==1.1.1
django-contrib-comments==1.9.2
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-filter==2.2.0
django-gravatar2==1.4.3
django-guardian==2.2.0
django-pgviews==0.5.7
djangorestframework==3.11.0
entrypoints==0.3
fits2image==0.4.3
gevent==1.4.0
graypy==2.1.0
greenlet==0.4.15
gunicorn==20.0.4
gwcs==0.13.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
keyring==19.2.0
kiwisolver==1.1.0
libmagic==1.0
llvmlite==0.31.0
lxml==4.5.0
MarkupSafe==1.1.1
mechanize==0.4.3
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numba==0.47.0
numpy==1.18.2
oauthlib==3.1.0
pandas==1.1.5
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.0
plotly==4.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.5
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyqtgraph==0.10.0
pyrsistent==0.15.5
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scipy==1.3.2
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
specutils==1.0
sqlparse==0.3.0
suds-py3==1.4.0.0
terminado==0.8.3
testpath==0.4.4
tom-antares==0.1.0
tom-lt==0.4.0
tomtoolkit==1.5.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==4.1.4
zipp==0.6.0
django-recaptcha==2.0.6